backend/requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.5.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart>=0.0.7
python-dotenv